refactor: move `loadAuthValues` to `~/services/Tools/credentials`
